chore(web): move shortcut util inside `/shortcut`
