+++ packages/web/src/common/utils/shortcut/shortcut.util.tsx
