+++ packages/web/src/views/Forms/EventForm/MigrateBelowMenuButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { ArrowDown } from "@phosphor-icons/react";
-import { getMetaKey } from "@web/common/utils/shortcut.util";
+import { getMetaKey } from "@web/common/utils/shortcut/shortcut.util";
 import { Text } from "@web/components/Text";
 import MenuItem from "../ActionsMenu/MenuItem";
