+++ packages/web/src/views/Forms/EventForm/SaveSection/SaveSection.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from "react";
 import { Priority } from "@core/constants/core.constants";
-import { getMetaKey } from "@web/common/utils/shortcut.util";
+import { getMetaKey } from "@web/common/utils/shortcut/shortcut.util";
 import { Btn, StyledSaveBtn } from "@web/components/Button/styled";
 import { Text } from "@web/components/Text";
 import { TooltipWrapper } from "@web/components/Tooltip/TooltipWrapper";
